# Scripts for decoding and encoding using numba for performance gain