--- a/CONTRIBUTING.md
+++ b/CONTRIBUTING.md
@@ -8,3 +8,4 @@ Contributors
 
  * [Will McGinnis](will@pedalwrencher.com)
  * [Leonard Norrgard](leonard.norrgard@gmail.com)
+ * [Ilyas Moutawwakil](ilyas.moutawwakil@gmail.com)