--- a/pygeohash/__init__.py
+++ b/pygeohash/__init__.py
@@ -11,6 +11,7 @@
 from .distances import geohash_approximate_distance, geohash_haversine_distance
 from .geohash import encode, decode, decode_exactly
 from .stats import mean, northern, southern, eastern, western, variance, std
+from .nbgeohash import nb_decode_exactly, nb_point_encode, nb_point_decode, nb_vector_encode, nb_vector_decode
 
 __author__ = 'willmcginnis'
 
@@ -20,6 +21,11 @@ __all__ = [
     'encode',
     'decode',
     'decode_exactly',
+    'nb_point_encode',
+    'nb_point_decode',
+    'nb_vector_encode',
+    'nb_vector_decode',
+    'nb_decode_exactly',
     'mean',
     'northern',
     'southern',