--- a/pygeohash/__init__.py
+++ b/pygeohash/__init__.py
@@ -11,7 +11,6 @@
 from .distances import geohash_approximate_distance, geohash_haversine_distance
 from .geohash import encode, decode, decode_exactly
 from .stats import mean, northern, southern, eastern, western, variance, std
-from .nbgeohash import nb_decode_exactly, nb_point_encode, nb_point_decode, nb_vector_encode, nb_vector_decode
 
 __author__ = 'willmcginnis'
 
@@ -21,11 +20,6 @@ __all__ = [
     'encode',
     'decode',
     'decode_exactly',
-    'nb_point_encode',
-    'nb_point_decode',
-    'nb_vector_encode',
-    'nb_vector_decode',
-    'nb_decode_exactly',
     'mean',
     'northern',
     'southern',
@@ -34,3 +28,19 @@ __all__ = [
     'variance',
     'std'
 ]
+
+try:
+    # Soft dependency
+    import numpy, numba
+    from .nbgeohash import nb_decode_exactly, nb_encode, nb_decode, nb_vector_encode, nb_vector_decode
+    __all__ += [
+        'nb_encode',
+        'nb_decode',
+        'nb_vector_encode',
+        'nb_vector_decode',
+        'nb_decode_exactly'
+    ]
+
+except ImportError:
+    import logging
+    logging.warn(f"Numpy and Numba are soft dependencies to use the numba geohashing functions. \nCan only import pure python functions")