--- a/pygeohash/__init__.py
+++ b/pygeohash/__init__.py
@@ -43,4 +43,4 @@ try:
 
 except ImportError:
     import logging
-    logging.warn(f"Numpy and Numba are soft dependencies to use the numba geohashing functions. \nCan only import pure python functions")
+    logging.warn(f"Numpy and Numba are soft dependencies to use the numba geohashing functions. \nCan only import/use native python functions.")