--- a/pygeohash/nbgeohash.py
+++ b/pygeohash/nbgeohash.py
@@ -3,7 +3,7 @@
    :platform: Unix, Windows
    :synopsis: A module for encoding to and decoding from the geohash system
 
-.. moduleauthor:: ilyas Moutawwakil <moutawwakil.ilyas.tsi@gmail.com>
+.. moduleauthor:: ilyas Moutawwakil <ilyas.moutawwakil@gmail.com>
 
 
 """
@@ -11,7 +11,7 @@
 
 from math import log10
 
-__base32 = '0123456789bcdefghjkmnpqrstuvwxyz'
+__base32 = "0123456789bcdefghjkmnpqrstuvwxyz"
 #  Note: the alphabet in geohash differs from the common base32 alphabet described in IETF's RFC 4648
 # (http://tools.ietf.org/html/rfc4648)
 
@@ -20,11 +20,12 @@ try:
     import numpy as np
     from numba import njit, types
 except ImportError:
-    print('Numpy and Numba are soft dependencies to use this feature.')
+    print("Numpy and Numba are soft dependencies to use this feature.")
     raise ImportError(
-        "Couldn't import numpy or numba, make sure they are installed properly.")
+        "Couldn't import numpy or numba, make sure they are installed properly."
+    )
 
-__author__ = 'Ilyas Moutawwakil'
+__author__ = "Ilyas Moutawwakil"
 
 
 @njit(cache=True, fastmath=True)
@@ -34,6 +35,7 @@ def base32_to_int(s: types.char) -> types.uint8:
     Surprisingly, on Numba this approach is faster than all the others.
     More specifically, because the numba hashtable (dictionary) is slower.
     """
+    assert s in __base32
     res = ord(s) - 48
     if res > 9:
         res -= 40
@@ -56,29 +58,29 @@ def nb_decode_exactly(geohash: types.string) -> types.Tuple:
     number).
     """
 
-    lat_interval_neg, lat_interval_pos, lon_interval_neg, lon_interval_pos = -90, 90, -180, 180
+    lat_interval_neg, lat_interval_pos, lon_interval_neg, lon_interval_pos = (
+        -90,
+        90,
+        -180,
+        180,
+    )
     lat_err, lon_err = 90, 180
-    masks = np.array([16, 8, 4, 2, 1])
     is_even = True
     for c in geohash:
         cd = base32_to_int(c)
-        for mask in masks:
+        for mask in (16, 8, 4, 2, 1):
             if is_even:  # adds longitude info
                 lon_err /= 2
                 if cd & mask:
-                    lon_interval_neg = (lon_interval_neg +
-                                        lon_interval_pos) / 2
+                    lon_interval_neg = (lon_interval_neg + lon_interval_pos) / 2
                 else:
-                    lon_interval_pos = (lon_interval_neg +
-                                        lon_interval_pos) / 2
+                    lon_interval_pos = (lon_interval_neg + lon_interval_pos) / 2
             else:  # adds latitude info
                 lat_err /= 2
                 if cd & mask:
-                    lat_interval_neg = (lat_interval_neg +
-                                        lat_interval_pos) / 2
+                    lat_interval_neg = (lat_interval_neg + lat_interval_pos) / 2
                 else:
-                    lat_interval_pos = (lat_interval_neg +
-                                        lat_interval_pos) / 2
+                    lat_interval_pos = (lat_interval_neg + lat_interval_pos) / 2
             is_even = not is_even
     lat = (lat_interval_neg + lat_interval_pos) / 2
     lon = (lon_interval_neg + lon_interval_pos) / 2
@@ -101,7 +103,7 @@ def nb_point_decode(geohash: types.string) -> types.Tuple:
     return lat, lon
 
 
-@njit(fastmath=True, parallel=True)
+@njit(fastmath=True)
 def nb_vector_decode(geohashes: types.Array) -> types.Tuple:
     """
     Decode geohashes, returning two Arrays of floats with latitudes and longitudes containing only relevant digits.
@@ -123,12 +125,19 @@ def nb_vector_decode(geohashes: types.Array) -> types.Tuple:
 
 
 @njit(fastmath=True)
-def nb_point_encode(latitude: types.float64, longitude: types.float64, precision: types.int8 = 12) -> types.string:
+def nb_point_encode(
+    latitude: types.float64, longitude: types.float64, precision: types.int8 = 12
+) -> types.string:
     """
     Encode a point given by latitude and longitude to a geohash of the specified precision.
     """
-    lat_interval_neg, lat_interval_pos, lon_interval_neg, lon_interval_pos = -90, 90, -180, 180
-    geohash = np.zeros(precision, dtype='<U1')
+    lat_interval_neg, lat_interval_pos, lon_interval_neg, lon_interval_pos = (
+        -90,
+        90,
+        -180,
+        180,
+    )
+    geohash = np.zeros(precision, dtype="<U1")
     bits = np.array([16, 8, 4, 2, 1])
     bit = 0
     ch = 0
@@ -159,17 +168,19 @@ def nb_point_encode(latitude: types.float64, longitude: types.float64, precision
             ch = 0
             n += 1
 
-    return ''.join(geohash)
+    return "".join(geohash)
 
 
-@njit(fastmath=True, parallel=True)
-def nb_vector_encode(latitudes: types.Array, longitudes: types.Array, precision: types.int8 = 12) -> types.Array:
+@njit(fastmath=True)
+def nb_vector_encode(
+    latitudes: types.Array, longitudes: types.Array, precision: types.int8 = 12
+) -> types.Array:
     """
     Encode a vector of points given by latitudes and longitudes to a vector of geohashes of the specified precision.
     This is not exactly a vectorized version of nb_point_encode, but it is way faster and gets faster as the number of points increase.
     """
     n = len(latitudes)
-    geohashes = np.empty(n, dtype='<U12')
+    geohashes = np.empty(n, dtype="<U12")
     for i in range(n):
         geohashes[i] = nb_point_encode(latitudes[i], longitudes[i], precision)
     return geohashes