Copied from what I did in [geohash-on-steroids](https://github.com/IlyasMoutawwakil/geohash-on-steroids):
- Solves #8 
## Numba geohash
Scripts to encode to and decode from the geohash system in a python-optimized way.

## Dependencies
Optimized functions are created with the `njit` decorators and using arrays so the only dependencies are [Numba](https://github.com/numba/numba) and [Numpy](https://github.com/numpy/numpy).

## Performance
As you can see in my [notebook](https://github.com/IlyasMoutawwakil/geohash-on-steroids/blob/main/performance_tests.ipynb), performance gain in comparison to the default python package [pygeohash](https://github.com/wdm0006/pygeohash) is the following:

```python
%%timeit
point_decode(geohash) # pygeohash
# Output: 20.4 µs ± 367 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

```python
%%timeit
nb_point_decode(geohash) # nbgeohash
# Output: 4.48 µs ± 16.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

```python
%%timeit
point_encode(latitude, longitude) # pygeohash
# Output: 92.8 µs ± 2.37 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

```python
%%timeit
nb_point_encode(latitude, longitude) # nbgeohash
# Output: 11.2 µs ± 663 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```
But geohashing is generally performaded on large amounts of data points so I made a vector-wise implimentation that perform well at large scale:

```python
%%timeit
np_vector_decode(geohashes) # a numpy vectorization of pygeohash's decode function
# Output: 2.09 s ± 25.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

```python
%%timeit
nb_vector_decode(geohashes) # nbgeohash
# Output: 164 ms ± 1.66 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

```python
%%timeit
np_vector_encode(latitudes, longitudes) # a numpy vectorization of pygeohash's encode function
# Output: 2.57 s ± 53.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

```python
%%timeit
nb_vector_encode(latitudes, longitudes) # nbgeohash
# Output: 443 ms ± 12.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```