@wdm0006 2021-11-24
I think keeping the new tests in a separate file (one for vectorized and one for the numba versions) would be cleanest.