@IlyasMoutawwakil 2021-11-25
There's some fails in distance and stats, but I guess they're just python rounding errors
Anyway, here's the unittest log:
```
test_check_validity (tests.test_geohash.TestGeohash) ... ok  
test_decode (tests.test_geohash.TestGeohash) ... ok
test_distance (tests.test_geohash.TestGeohash) ... FAIL      
test_encode (tests.test_geohash.TestGeohash) ... ok
test_stats (tests.test_geohash.TestGeohash) ... FAIL
test_decode (tests.test_nbgeohash.TestNumbaPointGeohash) ... ok
test_encode (tests.test_nbgeohash.TestNumbaPointGeohash) ... ok
test_decode (tests.test_nbgeohash.TestNumbaVectorGeohash) ... ok
test_encode (tests.test_nbgeohash.TestNumbaVectorGeohash) ... ok
```