@IlyasMoutawwakil 2022-01-24
> Overall looks good but let's use almostEquals to avoid rounding error issues in the tests and make sure that the soft dependencies are actually optional.

actually there's no rounding issue in decoding or coding to geohash, the issue is in distance and stats (which i did'nt implement), I can change them but that should be in a another PR ?